+++ src/app/advanced-examples/minesweeper/minesweeper.ts
@@ -16,9 +16,9 @@ export const Minesweeper = () => {
   );
 
   return Div(
-    GameBoard(board, dispatch),
-    Controls(startTime, endTime, gameStage, dispatch),
+    GameBoard({ board, dispatch }),
+    Controls({ startTime, endTime, gameStage, dispatch }),
   ).pipe(
-    classes('minesweeper'),
+    classes`minesweeper`,
   );
 };