+++ src/app/basic-examples/state-and-events.ts
@@ -5,6 +5,7 @@ export const ClickCounter = () => {
   const clicks = new BehaviorSubject(0);
 
   return Button`Clicks: ${clicks}`.pipe(
-    event('click', () => clicks.next(clicks.value + 1)),
+    event.click(() => clicks.next(clicks.value + 1)),
+    // May also be written as: `event('click', () => clicks.next(clicks.value + 1))`,
   );
 };