+++ src/lib/rxfm/attributes/attributes.ts
@@ -1,8 +1,8 @@
-import { Observable } from "rxjs";
-import { distinctUntilChanged, startWith, tap } from "rxjs/operators";
+import { combineLatest, Observable, of } from "rxjs";
+import { distinctUntilChanged, map, startWith, tap } from "rxjs/operators";
 import { Component, componentOperator, ComponentOperator, ElementType } from "../components";
 import { operatorIsolationService } from "../operator-isolation-service";
-import { coerceToObservable, PartialRecord, TypeOrObservable } from "../utils";
+import { coerceToObservable, PartialRecord, StringLike, TypeOrObservable } from "../utils";
 import { AttributeMetadataDictionary, AttributeMetadataObject, setAttributes } from "./attribute-operator-isolation";
 import { ClassType } from "./classes";
 import { HTMLAttributes } from "./html";
@@ -50,13 +50,11 @@ const setAttribute = (element: ElementType, key: string, val: string | null) => 
   }
 };
 
-/**
- * An observable operator to manage an attribute on an RxFM component.
- * @param type The attribute type.
- * @param value The attribute value or an observable emitting the value.
- * @param externalSymbol Implementation detail so that this operator may be used as the basis for the attributes operator.
- */
-export function attribute<T extends ElementType>(
+type BasicAttributeOperator = {
+  <T extends ElementType>(type: string, value?: TypeOrObservable<AttributeType>): ComponentOperator<T>;
+};
+
+function basicAttributeOperator<T extends ElementType>(
   type: string,
   value: TypeOrObservable<AttributeType> = '',
   externalSymbol?: symbol,
@@ -78,6 +76,54 @@ export function attribute<T extends ElementType>(
     );
   });
 }
+
+type IndividualAttributeOperator = {
+  <T extends ElementType>(value: TypeOrObservable<AttributeType>): ComponentOperator<T>;
+  // TODO: Replace return type with ComponentOperator once TS tagged template operator type inference is fixed.
+  (templateStrings: TemplateStringsArray, ...expressions: TypeOrObservable<StringLike>[]):
+    <T extends ElementType>(component: Component<T>) => Component<T>;
+};
+
+/**
+ * Create an attribute operator for an individual attribute key.
+ * @param attributeOperator The generic attribute operator.
+ * @param key The attribute type key.
+ */
+ function getIndividualAttributeOperator(attributeOperator: BasicAttributeOperator, key: AttributeKeys): IndividualAttributeOperator {
+  return <T extends ElementType>(
+    valueOrTemplateStrings: TypeOrObservable<AttributeType> | TemplateStringsArray,
+    ...expressions: TypeOrObservable<StringLike>[]
+  ): ComponentOperator<T> => {
+    if (!Array.isArray(valueOrTemplateStrings)) {
+      return attributeOperator(key, valueOrTemplateStrings as TypeOrObservable<AttributeType>);
+    } else {
+      const attributeObservables = (valueOrTemplateStrings as TemplateStringsArray)
+        .reduce<Observable<StringLike>[]>((acc, str, i) => {
+          acc.push(of(str));
+          if (expressions[i]) acc.push(coerceToObservable(expressions[i]));
+          return acc;
+        }, []);
+      const attributeObservable = combineLatest(attributeObservables).pipe(
+        map(strings => strings.join('')),
+      );
+      return attributeOperator(key, attributeObservable);
+    }
+  };
+}
+
+type AttributeOperators = {
+  [E in AttributeKeys]: IndividualAttributeOperator;
+};
+
+export type AttributeOperator = BasicAttributeOperator & AttributeOperators;
+
+/**
+ * An observable operator to manage an attribute on an RxFM component.
+ * Alternatively attribute operators for specific attribute types may be accessed directly as properties eg: `attribute.id('app')`.
+ * @param type The attribute type.
+ * @param value The attribute value or an observable emitting the value.
+ */
+export const attribute = new Proxy(basicAttributeOperator, { get: getIndividualAttributeOperator }) as AttributeOperator;
 
 /**
  * Element attributes which have a different interface to others.
@@ -130,9 +176,9 @@ export function attributes<T extends ElementType>(
       const symbol = Symbol('Attributes Operator');
       return Object.keys(attributesDict).reduce((component, key) => {
         return component.pipe(
-          attribute(key, attributesDict[key as AttributeKeys], symbol),
+          basicAttributeOperator(key, attributesDict[key as AttributeKeys], symbol),
         );
       }, input);
-    }
+    };
   }
 }