+++ src/lib/rxfm/attributes/styles.ts
@@ -1,8 +1,8 @@
-import { Observable } from "rxjs";
+import { combineLatest, Observable, of } from "rxjs";
 import { distinctUntilChanged, map, startWith, tap } from "rxjs/operators";
 import { Component, componentOperator, ComponentOperator, ElementType } from "../components";
 import { operatorIsolationService } from "../operator-isolation-service";
-import { coerceToObservable, NullLike, TypeOrObservable } from "../utils";
+import { coerceToObservable, NullLike, StringLike, TypeOrObservable } from "../utils";
 import { AttributeMetadataDictionary, AttributeMetadataObject, setAttributes } from "./attribute-operator-isolation";
 
 /**
@@ -11,7 +11,7 @@ import { AttributeMetadataDictionary, AttributeMetadataObject, setAttributes } f
 export type StyleKeys = Extract<
   keyof { [K in keyof CSSStyleDeclaration as CSSStyleDeclaration[K] extends string ? K : never]: CSSStyleDeclaration[K] },
   string
->
+>;
 
 /**
  * The types which may be applied as a style.
@@ -37,13 +37,11 @@ const setStyle = (element: ElementType, key: StyleKeys, value: string | null) =>
   }
 };
 
-/**
- * An observable operator to manage a style on an RxFM component.
- * @param name The style name.
- * @param value The style value or an observable emitting the value.
- * @param externalSymbol Implementation detail so that this operator may be used as the basis for the styles operator.
- */
-export function style<T extends ElementType, K extends StyleKeys>(
+type BasicStyleOperator = {
+  <T extends ElementType, K extends StyleKeys>(name: K, value: TypeOrObservable<StyleType>): ComponentOperator<T>;
+};
+
+function basicStyleOperator<T extends ElementType, K extends StyleKeys>(
   name: K,
   value: TypeOrObservable<StyleType>,
   externalSymbol?: symbol,
@@ -66,6 +64,54 @@ export function style<T extends ElementType, K extends StyleKeys>(
     );
   });
 }
+
+type IndividualStyleOperator = {
+  <T extends ElementType>(value: TypeOrObservable<StyleType>): ComponentOperator<T>;
+  // TODO: Replace return type with ComponentOperator once TS tagged template operator type inference is fixed.
+  (templateStrings: TemplateStringsArray, ...expressions: TypeOrObservable<StringLike>[]):
+    <T extends ElementType>(component: Component<T>) => Component<T>;
+};
+
+/**
+ * Create a style operator for an individual style key.
+ * @param styleOperator The generic style operator.
+ * @param key The style type key.
+ */
+ function getIndividualStyleOperator(styleOperator: BasicStyleOperator, key: StyleKeys): IndividualStyleOperator {
+  return <T extends ElementType>(
+    valueOrTemplateStrings: TypeOrObservable<StyleType> | TemplateStringsArray,
+    ...expressions: TypeOrObservable<StringLike>[]
+  ): ComponentOperator<T> => {
+    if (!Array.isArray(valueOrTemplateStrings)) {
+      return styleOperator(key, valueOrTemplateStrings as TypeOrObservable<StyleType>);
+    } else {
+      const styleObservables = (valueOrTemplateStrings as TemplateStringsArray)
+        .reduce<Observable<StringLike>[]>((acc, str, i) => {
+          acc.push(of(str));
+          if (expressions[i]) acc.push(coerceToObservable(expressions[i]));
+          return acc;
+        }, []);
+      const styleObservable = combineLatest(styleObservables).pipe(
+        map(strings => strings.join('')),
+      );
+      return styleOperator(key, styleObservable);
+    }
+  };
+}
+
+type StyleOperators = {
+  [E in StyleKeys]: IndividualStyleOperator;
+};
+
+export type StyleOperator = BasicStyleOperator & StyleOperators;
+
+/**
+ * An observable operator to manage a style on an RxFM component.
+ * Alternatively style operators for specific style types may be accessed directly as properties eg: `style.color('red')`.
+ * @param name The style name.
+ * @param value The style value or an observable emitting the value.
+ */
+export const style = new Proxy(basicStyleOperator, { get: getIndividualStyleOperator }) as StyleOperator;
 
 /**
  * A dictionary of styles or observable styles to be used in the 'styles' operator.
@@ -109,9 +155,9 @@ export function styles<T extends ElementType>(
       const symbol = Symbol('Styles Operator');
       return Object.keys(stylesDict).reduce((component, key) => {
         return component.pipe(
-          style(key as StyleKeys, stylesDict[key as StyleKeys], symbol),
+          basicStyleOperator(key as StyleKeys, stylesDict[key as StyleKeys], symbol),
         );
       }, input);
-    }
+    };
   }
 }