+++ src/lib/rxfm/utils/utils.ts
@@ -67,8 +67,8 @@ export type DestructuredObservable<T> = {
  */
 export function destructure<T> (source: Observable<T>, share = true): DestructuredObservable<T> {
   const sharedSource = share ? reuse(source) : source;
-  const handler = {
-    get: (_: DestructuredObservable<T>, prop: string | symbol) => selectFrom(sharedSource, prop as keyof T),
+  const handler: ProxyHandler<DestructuredObservable<T>> = {
+    get: (_, prop) => selectFrom(sharedSource, prop as keyof T),
   };
   return new Proxy({} as DestructuredObservable<T>, handler);
 }
@@ -103,9 +103,31 @@ export function using<T, U>(source: Observable<T>, action: (value: T) => U): Obs
 }
 
 /**
+ * Access a property on an object using an observable emitting object keys.
+ * This is equivalent to: `key.pipe(map(k => value[k]))` but will only emit distinct values.
+ * @param value An object of type T.
+ * @param key A key of T (K) observable.
+ * @returns An observable emitting T[K].
+ */
+export function access<T, K extends keyof T>(value: T, key: Observable<K>): Observable<T[K]> {
+  return key.pipe(
+    distinctUntilChanged(),
+    map(k => value[k]),
+    distinctUntilChanged(),
+  );
+}
+
+export interface ConditionalOptions<T, S, F = undefined> {
+  if: Observable<T>,
+  then: S | Observable<S>,
+  else: F | Observable<F>,
+}
+
+/**
  * A function taking a source observable and either emitting the success value or the fail value depending on whether
  * the source emits a truthy or falsy value.
  * @param source An observable of type T.
+ * Alternatively this may be an object containing the three arguments as properties, using the keys `if`, `then`, and `else`.
  * @param successValue The value of type S (or observable emitting type S) to return if T is truthy.
  * @param failValue The value of type F (or observable emitting type F) to return if T is falsy.
  * @returns Returns an observable emitting either the success value of type S or the fail value of type F depending on the
@@ -115,10 +137,22 @@ export function conditional<T, S, F = undefined>(
   source: Observable<T>,
   successValue: S | Observable<S>,
   failValue?: F | Observable<F>,
+): Observable<S | F>;
+export function conditional<T, S, F = undefined>(options: ConditionalOptions<T, S, F>): Observable<S | F>;
+export function conditional<T, S, F = undefined>(
+  sourceOrOptions: Observable<T> | ConditionalOptions<T, S, F>,
+  successValue?: S | Observable<S>,
+  failValue?: F | Observable<F>,
 ): Observable<S | F> {
+  const { if: source, then: thenVal, else: elseVal } = sourceOrOptions instanceof Observable ? {
+    if: sourceOrOptions,
+    then: successValue as S | Observable<S>,
+    else: failValue,
+  }: sourceOrOptions;
+
   return source.pipe(
     distinctUntilChanged(),
-    switchMap(value => value ? coerceToObservable(successValue) : coerceToObservable(failValue as F)),
+    switchMap(value => coerceToObservable(value ? thenVal : elseVal as F | Observable<F>)),
     distinctUntilChanged(),
   );
 }
@@ -138,6 +172,17 @@ export function reuse<T>(source: Observable<T>): Observable<T> {
 }
 
 /**
+ * @returns An observable emitting the logical NOT value of the source observable's emissions.
+ */
+ export function notGate(source: Observable<any>): Observable<boolean> {
+  return source.pipe(
+    distinctUntilChanged(),
+    map(val => !val),
+    distinctUntilChanged(),
+  );
+}
+
+/**
  * Take a spread array of observables and emit the logical AND value of all of their emissions whenever it changes.
  */
 export function andGate(...sources: Observable<any>[]): Observable<boolean> {
@@ -147,6 +192,13 @@ export function andGate(...sources: Observable<any>[]): Observable<boolean> {
     map(values => values.every(value => Boolean(value))),
     distinctUntilChanged(),
   );
+}
+
+/**
+ * Take a spread array of observables and emit the logical NAND value of all of their emissions whenever it changes.
+ */
+export function nandGate(...sources: Observable<any>[]): Observable<boolean> {
+  return notGate(andGate(...sources));
 }
 
 /**
@@ -162,14 +214,10 @@ export function orGate(...sources: Observable<any>[]): Observable<boolean> {
 }
 
 /**
- * @returns An observable emitting the logical NOT value of the source observable's emissions.
+ * Take a spread array of observables and emit the logical NOR value of all of their emissions whenever it changes.
  */
-export function notGate(source: Observable<any>): Observable<boolean> {
-  return source.pipe(
-    distinctUntilChanged(),
-    map(val => !val),
-    distinctUntilChanged(),
-  );
+export function norGate(...sources: Observable<any>[]): Observable<boolean> {
+  return notGate(orGate(...sources));
 }
 
 /**